Clean up docs generation and get ready to make this deployable
