+++ docs/next.config.js
@@ -1,17 +1,5 @@
-const withMdxEnhanced = require('next-mdx-enhanced');
-
-module.exports = withMdxEnhanced({
-  layoutPath: 'src/layouts',
-  defaultLayout: true,
-  fileExtensions: ['mdx'],
-  remarkPlugins: [
-    require('remark-emoji'),
-    require('remark-slug'),
-    require('remark-code-titles'),
-  ],
-})({
-  devIndicators: {
-    autoPrerender: false,
-  },
+module.exports = {
+  poweredByHeader: false,
   reactStrictMode: true,
-});
+  target: 'serverless',
+};
