--- docs/nodemon.json
@@ -1,11 +1,0 @@
-{
-  "verbose": false,
-  "ignore": ["node_modules", ".next"],
-  "watch": [
-    "../packages/cascara",
-    "next.config.js",
-    "package.json",
-    "yarn.lock"
-  ],
-  "ext": "js json"
-}
