+++ docs/package.json
@@ -1,29 +1,26 @@
 {
-  "name": "cascara-docs",
+  "name": "docs",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "start": "next-remote-watch ../packages/cascara",
     "build": "next build",
-    "serve": "next start"
+    "postbuild": "next-on-netlify"
   },
   "dependencies": {
-    "@espressive/cascara": "link:../packages/cascara",
-    "@espressive/legacy-css": "link:../utils/legacy-css",
-    "@mapbox/rehype-prism": "0.5.0",
+    "@espressive/cascara": "file:../packages/cascara",
+    "@espressive/legacy-css": "file:../utils/legacy-css",
     "directory-tree": "2.2.4",
     "framer-motion": "2.5.5",
     "gray-matter": "4.0.2",
     "next": "9.5.1",
-    "next-mdx-enhanced": "3.1.0",
     "next-mdx-remote": "1.0.0",
+    "next-remote-watch": "0.2.0",
     "prism-react-renderer": "1.1.1",
-    "react": "link:/node_modules/react",
-    "react-dom": "link:/node_modules/react-dom",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
     "react-live": "2.2.2",
-    "remark-code-extra": "1.0.1",
     "remark-code-frontmatter": "1.0.0",
-    "remark-code-titles": "0.1.1",
     "remark-emoji": "2.1.0",
     "remark-slug": "6.0.0",
     "semantic-ui-react": "1.2.0",
