--- docs/src/layouts/MDXLayout.js
@@ -1,19 +1,0 @@
-import Head from 'next/head';
-
-const MDXLayout = (frontMatter) => {
-  const Page = ({ children: content }) => {
-    return (
-      <>
-        <Head>
-          <title>{frontMatter.title || 'Cascara'}</title>
-          <link href='/favicon.ico' rel='icon' />
-        </Head>
-        <h1>{frontMatter.title}</h1>
-        {content}
-      </>
-    );
-  };
-
-  return Page;
-};
-export default MDXLayout;
