--- docs/src/pages/test.mdx
@@ -1,11 +1,0 @@
----
-title: Test Title
----
-
-# Testing MDX Files
-
-Okay let's see how this works.
-
-```jsx
-const Component = (props) => {...}
-```
