+++ netlify.toml
@@ -2,14 +2,10 @@
 # unless otherwise overridden by more specific contexts.
 [build]
 
-  # Directory that contains the deploy-ready HTML files and assets generated by
-  # the build. This is relative to the base directory if one has been set, or the
-  # root directory if a base has not been set. This sample publishes the
-  # directory located at the absolute path "root/project/build-output"
-  publish = "build/docs"
-
   # Default build command.
-  command = "yarn workspace button run build && yarn docz:build"
+  command = "yarn cascara build && yarn docs build"
+  functions = "out_functions"
+  publish   = "out_publish"
 
   # Directory with the serverless Lambda functions to deploy to AWS.
   # functions = "functions/"
