+++ packages/cascara/README.md
@@ -1,0 +1,9 @@
+# @espressive/cascara
+
+> NOTE: This will eventually get replaced with a generated README which will probably link to the cascara.design docs.
+
+If you happen to be here ahead of schedule:
+
+```javascript
+import { Some, Cool, Components } from '@espressive/cascara';
+```
