--- packages/cascara/src/ui/README.md
@@ -1,7 +1,0 @@
-# UI
-
-All of our components here should initially be our UI elements that the design system is built from. The API surface area of all components should be as simple as possible to make use easier, but also to make sure we are keeping consistency in our UI.
-
-In the long run, this area might need to be organized into different areas if the system grows so we can take care of all of our needs.
-
-import { Button } from 'cascara'
